+++ src/main/java/de/johni0702/minecraft/bobby/ChunkSerializer.java
@@ -4,6 +4,7 @@ import com.google.common.hash.Hashing;
 import com.mojang.serialization.Codec;
 import de.johni0702.minecraft.bobby.ext.ChunkLightProviderExt;
 import de.johni0702.minecraft.bobby.ext.LightingProviderExt;
+import de.johni0702.minecraft.bobby.ext.WorldChunkExt;
 import net.minecraft.SharedConstants;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
@@ -13,6 +14,7 @@ import net.minecraft.nbt.NbtElement;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.nbt.NbtLongArray;
 import net.minecraft.nbt.NbtOps;
+import net.minecraft.network.packet.s2c.play.LightData;
 import net.minecraft.registry.DynamicRegistryManager;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKeys;
@@ -39,6 +41,7 @@ import org.apache.logging.log4j.Logger;
 import java.util.Arrays;
 import java.util.BitSet;
 import java.util.EnumSet;
+import java.util.Iterator;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Supplier;
@@ -342,9 +345,27 @@ public class ChunkSerializer {
         ChunkNibbleArray[] blockLight = new ChunkNibbleArray[chunkSections.length + 2];
         ChunkNibbleArray[] skyLight = new ChunkNibbleArray[chunkSections.length + 2];
         LightingProvider lightingProvider = world.getChunkManager().getLightingProvider();
-        for (int y = lightingProvider.getBottomY(), i = 0; y < lightingProvider.getTopY(); y++, i++) {
-            blockLight[i] = lightingProvider.get(LightType.BLOCK).getLightSection(ChunkSectionPos.from(chunkPos, y));
-            skyLight[i] = lightingProvider.get(LightType.SKY).getLightSection(ChunkSectionPos.from(chunkPos, y));
+        LightData initialLightData = WorldChunkExt.get(original).bobby_getInitialLightData();
+        if (initialLightData != null) {
+            Iterator<byte[]> blockNibbles = initialLightData.getBlockNibbles().iterator();
+            Iterator<byte[]> skyNibbles = initialLightData.getSkyNibbles().iterator();
+            for (int y = lightingProvider.getBottomY(), i = 0; y < lightingProvider.getTopY(); y++, i++) {
+                boolean hasBlockData = initialLightData.getInitedBlock().get(i);
+                boolean isBlockZero = initialLightData.getUninitedBlock().get(i);
+                if (hasBlockData || isBlockZero) {
+                    blockLight[i] = hasBlockData ? new ChunkNibbleArray(blockNibbles.next().clone()) : new ChunkNibbleArray();
+                }
+                boolean hasSkyData = initialLightData.getInitedSky().get(i);
+                boolean isSkyZero = initialLightData.getUninitedSky().get(i);
+                if (hasSkyData || isSkyZero) {
+                    skyLight[i] = hasSkyData ? new ChunkNibbleArray(skyNibbles.next().clone()) : new ChunkNibbleArray();
+                }
+            }
+        } else {
+            for (int y = lightingProvider.getBottomY(), i = 0; y < lightingProvider.getTopY(); y++, i++) {
+                blockLight[i] = lightingProvider.get(LightType.BLOCK).getLightSection(ChunkSectionPos.from(chunkPos, y));
+                skyLight[i] = lightingProvider.get(LightType.SKY).getLightSection(ChunkSectionPos.from(chunkPos, y));
+            }
         }
 
         FakeChunk fake = new FakeChunk(world, chunkPos, chunkSections);